Fixed errors in category save function and changed category page layout. closes #7
