+++ iBudget/CategoryViewController.swift
@@ -19,12 +19,12 @@ class CategoryViewController: UIViewController {
     
     @IBAction func addCategoryButtonTapped(_ sender: UIButton) {
         
-        performSegue(withIdentifier: addCategorySegue, sender: self)
+       //performSegue(withIdentifier: addCategorySegue, sender: self)
     }
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        loadCategories()
+        //loadCategories()
         categoriesTableView.dataSource = self
         
         //Uncomment later when categories need to be clickable
@@ -37,99 +37,27 @@ class CategoryViewController: UIViewController {
         
         return context
     }
-    
-    func saveAll(){
-        
-        AppDelegate.sharedAppDelegate.coreDataStack.saveContext()
-         
-        print("Saved")
-        //getUserInfo()
-        print("Loaded")
-
-    }
 
     func loadCategories(){
         //TODO: Use to load original categories for the first time
-        
+        categoriesArray = []
         let request1: NSFetchRequest<Categories> = Categories.fetchRequest()
             do {
                 let items = try getContext().fetch(request1)
-    
-                for item in items{
-                    let category0 = Categories(context: getContext())
-                    category0.name = item.name
-                    category0.type = item.type
-                    category0.id = item.id
-                    category0.icon = item.icon
-                    
-                    categoriesArray.append(category0)
-                }
+                print(items.count)
+                categoriesArray.append(contentsOf: items)
                 categoriesTableView.reloadData()
                 
     
             } catch {
                 print ("error fetching data: \(error)")
             }
-        
-        //TODO: First add categories from data storage, then, if empty, add the basic categories.
-        if(categoriesArray.isEmpty)
-        {
-            let category = Categories(context: getContext())
-            category.name = "Food"
-            //Set type as 0 because they are the predefined categories
-            category.type = 0
-            var id = UUID()
-            category.id = id
-            //TODO: Change icons
-            category.icon = "fork.knife"
-            categoriesArray.append(category)
-            
-            let category2 = Categories(context: getContext())
-            category2.name = "Housing"
-            //Set type as 0 because they are the predefined categories
-            category2.type = 0
-            id = UUID()
-            category2.id = id
-            //TODO: Change icons
-            category2.icon = "house"
-            categoriesArray.append(category2)
-            
-            let category3 = Categories(context: getContext())
-            category3.name = "Transportation"
-            //Set type as 0 because they are the predefined categories
-            category3.type = 0
-            id = UUID()
-            category3.id = id
-            //TODO: Change icons
-            category3.icon = "car"
-            categoriesArray.append(category3)
-            
-            let category4 = Categories(context: getContext())
-            category4.name = "Utilities"
-            //Set type as 0 because they are the predefined categories
-            category4.type = 0
-            id = UUID()
-            category4.id = id
-            //TODO: Change icons
-            category4.icon = "wrench.and.screwdriver"
-            categoriesArray.append(category4)
-            
-            let category5 = Categories(context: getContext())
-            category5.name = "Medical"
-            //Set type as 0 because they are the predefined categories
-            category5.type = 0
-            id = UUID()
-            category5.id = id
-            //TODO: Change icons
-            category5.icon = "cross"
-            categoriesArray.append(category5)
-            
-            self.saveAll()
-        }
     }
     
     override func viewDidAppear(_ animated: Bool) {
-        categoriesTableView.reloadData()
+        
+        //categoriesTableView.reloadData()
+        loadCategories()
     }
 }
